+++ Sources/PatchDay/Sites/Details/SiteDetailViewController.swift
@@ -151,9 +151,15 @@ class SiteDetailViewController: UIViewController, UIPickerViewDelegate, UIPicker
         shouldChangeCharactersIn range: NSRange,
         replacementString string: String
     ) -> Bool {
-        MaxString.canSet(
-            currentString: textField.text ?? "", replacementString: string, range: range
+        let currentText = textField.text ?? ""
+        let result = TextFieldHelper.canSet(
+            currentString: currentText, replacementString: string, range: range
         )
+        if result.canReplace {
+            viewModel.selections.selectedSiteName = result.updatedText
+            enableSave()
+        }
+        return result.canReplace
     }
 
     @objc func saveButtonTapped(_ sender: Any) {